Update convert_column_types.R
Important update:
1. Move the `switch(file_ext, read_sas..` from `init(,server=` to this function, to allow make more manipulations to uploaded data;
2. Currently, some clinical modules requires the `arm_var` has non-missing values(not empty, NA); 
There isn't a good way to solve this problem. For now, modify the empty ACTARM as a specific value("Acutal ARM is Null") to indicate ACTARM is null. This can ensure those modules load correctly, and by comparing it with other ARM variables, the actual data situation can still be understood.